refactor(routing): harden max_token_id auto-config for robustness
Addresses several issues in the auto-configure max_token_id feature:

1. Integer overflow protection: Add bounds check when casting vocab_size
   to int32_t, clamping to INT32_MAX if exceeded (theoretical for now)

2. Observability: Log warning when vocab_size returns 0, indicating
   potential tokenizer loading issues

3. Config consistency: Update ShardedConfig after auto-configuring
   max_token_id so Application::local_config() returns correct value

4. Hot-reload preservation: Re-apply auto-config in reload_config()
   so max_token_id isn't reverted to file value on SIGHUP

Extracted logic into reusable apply_vocab_size_config() method.

+++ src/tokenizer_service.cpp
@@ -23,7 +23,7 @@ bool TokenizerService::is_loaded() const {
 
 size_t TokenizerService::vocab_size() const {
     if (!_impl) return 0;
-    return static_cast<size_t>(_impl->GetVocabSize());
+    return _impl->GetVocabSize();
 }
 
 } // namespace ranvier
